requirements.txt
fastapi
uvicorn
pyrebase4
firebase_admin
setuptools
google-generativeai